nvif: split out device interface definitions
